@@ -15,9 +15,9 @@
   "devDependencies": {
     "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tauri-apps/cli": "^2.0.0",
-    "svelte": "^4.2.0",
+    "svelte": "^5.0.0",
     "typescript": "^5.0.0",
-    "vite": "^6.0.0",
+    "vite": "^5.0.0",
     "svelte-check": "^4.0.0"
   }
 }
