@@ -1,6 +1,7 @@
+import { mount } from "svelte";
 import App from "./App.svelte";
 
-const app = new App({
+const app = mount(App, {
   target: document.getElementById("app")!,
 });
 
